+++ cmd/scribe/main.go
@@ -1,25 +1,37 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"net/http"
-
-	"github.com/kelseyhightower/envconfig"
+	"time"
 
 	"github.com/gorilla/mux"
+	"github.com/kelseyhightower/envconfig"
 	"github.com/payvision-development/scribe"
+	"github.com/payvision-development/scribe/freshservice"
+	"github.com/payvision-development/scribe/health"
 	"github.com/payvision-development/scribe/vss"
+	"github.com/urfave/negroni"
 )
 
 // Specification struct
 type Specification struct {
-	FreshserviceURL string `required:"true" split_words:"true"`
+	User               string `required:"true"`
+	Pass               string `required:"true"`
+	FreshserviceURL    string `required:"true" split_words:"true"`
+	FreshserviceEmail  string `required:"true" split_words:"true"`
+	FreshserviceApikey string `required:"true" split_words:"true"`
+	VstsApikey         string `required:"true" split_words:"true"`
 }
 
 var env Specification
 var events = make(chan *vss.Event, 10)
+
+var date = time.Now()
+var count uint32
 
 func main() {
 	fmt.Println("Scribe is alive")
@@ -31,20 +43,74 @@ func main() {
 
 	go eventRouter()
 
+	n := negroni.Classic()
 	router := mux.NewRouter()
-	router.HandleFunc("/vss-release", release).Methods("POST")
-	log.Fatal(http.ListenAndServe(":8000", router))
+
+	router.HandleFunc("/status", status).Methods("GET")
+
+	vssRouter := mux.NewRouter().PathPrefix("/vss").Subrouter().StrictSlash(true)
+	vssRouter.HandleFunc("/release", vssRelease).Methods("POST")
+
+	router.PathPrefix("/vss").Handler(negroni.New(
+		negroni.HandlerFunc(basicAuth),
+		negroni.Wrap(vssRouter),
+	))
+
+	n.UseHandler(router)
+	n.Run(":8080")
 }
 
-func release(rw http.ResponseWriter, req *http.Request) {
+func basicAuth(rw http.ResponseWriter, r *http.Request, next http.HandlerFunc) {
+	user, pass, _ := r.BasicAuth()
+
+	if env.User != user || env.Pass != pass {
+		rw.Header().Set("WWW-Authenticate", `Basic realm="Restricted"`)
+		http.Error(rw, "Unauthorized.", http.StatusUnauthorized)
+		return
+	}
+
+	next(rw, r)
+}
+
+func status(rw http.ResponseWriter, req *http.Request) {
+	status := health.Status{
+		Service:     "Scribe",
+		Description: "VSTS Release event integration with Freshservice",
+		Status:      "OK",
+		Version:     "1.1.0",
+		Info: health.Info{
+			Started: date,
+			Events:  count,
+		},
+	}
+
+	client := freshservice.NewClient(env.FreshserviceURL, env.FreshserviceEmail, env.FreshserviceApikey)
+
+	err := client.CheckEndpoint()
+	if err != nil {
+		status.Status = "KO: Failed to connect with Freshservice API endpoint"
+	}
+
+	json.NewEncoder(rw).Encode(status)
+}
+
+func vssRelease(rw http.ResponseWriter, req *http.Request) {
+	count++
 
 	b, err := ioutil.ReadAll(req.Body)
 	defer req.Body.Close()
+
 	if err != nil {
-		panic(err)
+		fmt.Println(err)
+		return
 	}
 
-	event := scribe.Parser(b)
+	event, err := scribe.Parser(b)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
 	events <- event
 }
 
@@ -52,15 +118,24 @@ func eventRouter() {
 	m := make(map[uint32]chan *vss.Event)
 
 	for event := range events {
-		release, ok := m[event.ReleaseTrackingCode]
+		d, ok := m[event.ReleaseTrackingCode]
 
 		if !ok {
-			release = make(chan *vss.Event)
-			m[event.ReleaseTrackingCode] = release
+			fmt.Printf("[Release: %v] Creating new session for release %v (%v)\n", event.ReleaseTrackingCode, event.ReleaseName, event.ReleaseID)
 
-			go scribe.Session(release)
+			d = make(chan *vss.Event)
+			m[event.ReleaseTrackingCode] = d
+
+			var v *vss.TFS
+			if event.ServerURL != "" || event.CollectionURL != "" {
+				v = vss.NewClient(event.ServerURL, event.CollectionURL, env.VstsApikey)
+			}
+
+			fs := freshservice.NewClient(env.FreshserviceURL, env.FreshserviceEmail, env.FreshserviceApikey)
+
+			go scribe.Session(event.ReleaseTrackingCode, d, fs, v)
 		}
 
-		release <- event
+		d <- event
 	}
 }